ecu-traq-webapp(2)/backend/requirements.txt

fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.2
